api/requirements.txt
Werkzeug==0.16.1
connexion == 1.1.15
python_dateutil == 2.6.0
setuptools >= 21.0.0
flask-cors
tensorflow==2.3.0
sklearn
psycopg2-binary 
numpy
nltk
